backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
APScheduler==3.10.4
ruamel.yaml==0.18.6
Werkzeug==3.0.3
pydantic==2.9.2
python-dateutil==2.9.0.post0
email-validator==2.2.0
